Fixed bug, CommandList.RenderTargets which was depended upon by XModelLoader via it's usage of PipelineState, was empty.  Introduced CommandList.SetRenderTargets to control these, and made a call to this on initialisation to ensure they are set prior to loading assets.  In order to retain  their CpuDescriptorHandles for use with ID3D12GraphicsCommandList.OMSetRenderTargets, I opted to introduce the RenderTargetView abstraction to store this together with the resource.
Fixed another bug, the Sampler abstraction was not actually making the call to CreateSampler!

Aligned root signature with that of HelloTexture, as that was clearly working.

For now, have remarked out the construction of mPipelineState, passing null in its place for the CommandList etc, as we instead use the pipeline states created for each model now.

Restored use of IASetIndexBuffer now that HelloTexture confirmed XModelLoader is working fine.

Remarked out the extra ConstantBufferViews that were inherited from DirectX12GameEngine materials but not used by HelloTexture to render the ship successfully.  This required corresponding changes in TextureShader.Accept to align them.

Refactored out mCommandQueue in favour of the mGraphicsDevice.DirectCommandQueue abstraction.

Cleaned up old code shader code no longer required.  mSrvHeap is present because HelloTexture used it, but it's not currently referenced in D3D12Mutiny because of the corresponding records in the GraphicsDevice class that are integrated with the ShaderGeneratorContext.

I tried making TextureShader.ColorTexture nullable however this interfered with shader compilation and so I reverted the change, but explicitly assigned null! in the constructor to avoid a warning.

Current state: no longer crashing, but the model isn't visible