Reduce transformations within the XModelLoader, leaving only a scale of 0.1f now.  Also, actually apply the FrameTransformMatrix (not doing this is why the model was skewed in earlier rendering).  - Compensated for this change in D3D12HelloTexture by using the vertex shader to further scale and transform since this sample doesn't have a camera, and we still want to actually see the ship.
XModelLoader now also returns the shader resource views, which was a precautionary change just in case the garbage collector was cleaning things up because we just had unmanaged code holding a reference to a descriptor, and possibly nothing olding the reference to the ComPtr in managed code after the PSO was constructed...

Minor typo in comment in HelloTriangle shader corrected.

D3D12Mutiny, more trial and error, application runs but can't see anything rendered.

Updated Vortice libraries to 3.2.0, and Pfim to 0.11.2.